change data types to avoid counter overflow
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/branches/classes@4610 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/workload.hpp
@@ -162,13 +162,14 @@ protected:
 
   int _channel_width;
 
-  int _region;
+  unsigned int _region;
 
-  int _count;
-  int _limit;
+  unsigned long long int _count;
+  unsigned long long int _limit;
 
-  int _scale;
-  int _skip;
+  unsigned long long int _scale;
+
+  unsigned long long int _skip;
 
   void _refill();
